Message: 11 通用组件：函数调用 confirm 组件

+++ src/views/layout/components/header/header-search/history.vue
@@ -34,6 +34,7 @@ const EMITS_ITEM_CLICK = 'itemClick'
 
 <script setup>
 import { useStore } from 'vuex'
+import { confirm } from '@/libs'
 
 defineOptions({
   name: 'history'
@@ -47,7 +48,13 @@ const store = useStore()
  * 删除所有记录
  */
 const onDeleteAllClick = () => {
-  store.commit('search/deleteAllHistory')
+  confirm('要删除所有历史记录吗？')
+    .then(() => {
+      store.commit('search/deleteAllHistory')
+    })
+    .catch((e) => {
+      console.log(e)
+    })
 }
 
 /**
